pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=2.3.0",
    "gunicorn",
    "psycopg>=3.2.12",
    "psycopg2>=2.9.11",
    "pytest>=9.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.4",
]

[project.scripts]
page-analyzer = "app:app"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest"
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
